Oppdaterte feil oppdaterSøknadMedBarn, den i BarnetilsynApp er korrekt. Gjort små endringer til consoleLogLokaltOgPreprod, endret navn på fil og fjernet implementeringen fra en del filer.
