+++ src/App.tsx
@@ -20,7 +20,6 @@ import { ESkjemanavn } from './utils/skjemanavn';
 import { useLokalIntlContext } from './context/LokalIntlContext';
 import { Alert, Loader } from '@navikt/ds-react';
 import { IBarn } from './models/steg/barn';
-import { consoleLogLokaltOgDev } from './utils/logLokaltOgDev';
 
 const App = () => {
   const [autentisert, settAutentisering] = useState<boolean>(false);
@@ -44,7 +43,6 @@ const App = () => {
   const fetchPersonData = () => {
     return hentPersonData()
       .then((response) => {
-        consoleLogLokaltOgDev(response, 'App.tsx response');
         settPerson({
           type: PersonActionTypes.HENT_PERSON,
           payload: response,
@@ -72,17 +70,11 @@ const App = () => {
 
   const oppdaterSøknadMedBarn = (person: IPerson, barneliste: IBarn[]) => {
     const barnMedLabels = oppdaterBarnMedLabel(barneliste, intl);
-    consoleLogLokaltOgDev(barnMedLabels, 'App.tsx barnMedLabels');
-    settSøknad((prevSøknad) => {
-      const prevBarn = prevSøknad.person.barn;
 
-      const oppdatertBarn = [...prevBarn, ...barnMedLabels];
-
-      return {
-        ...prevSøknad,
-        person: { ...person, barn: oppdatertBarn },
-      };
-    });
+    settSøknad((prevSøknad) => ({
+      ...prevSøknad,
+      person: { ...person, barn: barnMedLabels },
+    }));
   };
 
   const fetchToggles = () => {
