+++ src/søknad/steg/4-barnasbosted/OmAndreForelder.tsx
@@ -11,7 +11,7 @@ import { hentUid } from '../../../utils/autentiseringogvalidering/uuid';
 import { useLokalIntlContext } from '../../../context/LokalIntlContext';
 import IdentEllerFødselsdatoGruppe from '../../../components/gruppe/IdentEllerFødselsdatoGruppe';
 import { Checkbox, ErrorMessage, Textarea, TextField } from '@navikt/ds-react';
-import { consoleLogLokaltOgDev } from '../../../utils/logLokaltOgDev';
+import { consoleLogLokaltOgDev } from '../../../utils/consoleLogLokaltOgPreprod';
 import { slettIrrelevantPropertiesHvisHuketAvKanIkkeOppgiAnnenForelder } from '../../../helpers/steg/forelder';
 
 interface Props {
