--- src/utils/logLokaltOgDev.ts
@@ -1,8 +1,0 @@
-import Environment from '../Environment';
-
-export const consoleLogLokaltOgDev = (melding: any, sted?: string): void => {
-  const miljø = Environment().miljø;
-  if (miljø === 'local' || miljø === 'preprod') {
-    console.log(sted && `${sted}: `, melding);
-  }
-};
